requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.5.0
numpy>=1.19.2
pandas>=1.2.4
scikit-learn>=0.24.2
tqdm>=4.61.2
PyYAML>=5.4.1
pillow>=8.2.0
matplotlib>=3.4.2
seaborn>=0.11.1
pytest>=6.2.5 